Adding modules for routes & middlewares (user resource)
